collector.antiguos/datalab_essencials/requirements.txt
alabaster==0.7.10
alembic==0.9.5
anaconda-client==1.6.3
anaconda-navigator==1.6.4
anaconda-project==0.6.0
AnyQt==0.0.8
aplus==0.11.0
asn1crypto==0.22.0
astroid==1.4.9
astropy==1.3.2
astroquery==0.3.6.post2
autobahn==17.9.3
Babel==2.4.0
backports.functools-lru-cache==1.4
backports.shutil-get-terminal-size==1.0.0
bambi==0.1.0
beautifulsoup4==4.6.0
bitarray==0.8.1
blaze==0.10.1
bleach==2.0.0
bokeh==0.12.5
boto==2.46.1
Bottleneck==1.2.1
bqplot==0.10.0
cachetools==2.0.0
cassandra-driver==3.11.0
certifi==2017.7.27.1
cffi==1.10.0
chardet==3.0.4
click==6.7
cloudpickle==0.2.2
clyent==1.2.2
CMRESHandler==1.0.0
colorama==0.3.9
coloredlogs==7.3
conda==4.3.29
contextlib2==0.5.5
cryptography==1.8.1
cycler==0.10.0
Cython==0.25.2
cytoolz==0.8.2
dash==0.18.3
dash-core-components==0.12.5
dash-html-components==0.7.0
dash-renderer==0.9.0
dask==0.14.3
datashape==0.5.4
decorator==4.1.2
dill==0.2.7.1
distributed==1.16.3
docutils==0.13.1
elasticsearch==5.4.0
elasticsearch-dsl==5.3.0
entrypoints==0.2.3
et-xmlfile==1.0.1
fastcache==1.0.2
Flask==0.12.2
Flask-Compress==1.4.0
Flask-Cors==3.0.2
Flask-SeaSurf==0.2.2
freeopcua==0.90.3
future==0.16.0
futures==3.1.1
gevent==1.2.1
Ghost.py==0.2.3
glue-core==0.10.4
glue-vispy-viewers==0.7.2
glueviz==0.10.4
graphviz==0.8
greenlet==0.4.12
h5py==2.7.1
HeapDict==1.0.0
hide-code==0.4.0
html5lib==0.999999999
humanfriendly==4.4.1
idna==2.6
imagesize==0.7.1
ipykernel==4.6.1
ipython==5.3.0
ipython-genutils==0.2.0
ipyvolume==0.4.0
ipywebrtc==0.3.0
ipywidgets==7.0.1
isort==4.2.5
itsdangerous==0.24
jdcal==1.3
jedi==0.10.2
Jinja2==2.9.6
joblib==0.11
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.1.0
jupyter-console==5.1.0
jupyter-contrib-core==0.3.3
jupyter-contrib-nbextensions==0.3.1
jupyter-core==4.3.0
jupyter-dashboards==0.7.0
jupyter-highlight-selected-word==0.0.11
jupyter-latex-envs==1.3.8.2
jupyter-nbextensions-configurator==0.2.7
jupyterhub==0.7.2
jupyterlab==0.27.0
jupyterlab-launcher==0.5.3
kapteyn==2.3
keyring==9.3.1
lazy-object-proxy==1.2.2
linecache2==1.0.0
llvmlite==0.18.0
locket==0.2.0
lxml==3.7.3
Mako==1.0.7
MarkupSafe==1.0
matplotlib==2.1.0
matplotlib-venn==0.11.5
metakernel==0.20.7
mistune==0.7.4
mpi4py==2.0.0
mpmath==0.19
msgpack-python==0.4.8
multipledispatch==0.4.9
multiprocess==0.70.5
navigator-updater==0.1.0
nb-anacondacloud==1.4.0
nb-conda==2.2.1
nb-conda-kernels==2.1.0
nbbrowserpdf==0.2.1
nbconvert==4.3.0
nbformat==4.4.0
nbpresent==3.0.2
networkx==1.11
nltk==3.2.3
nose==1.3.7
notebook==5.1.0
numba==0.33.0
numexpr==2.6.2
numpy==1.12.1
numpydoc==0.6.0
oct2py==4.0.6
octave-kernel==0.27.0
odo==0.5.0
olefile==0.44
opcua-client==0.5.0
opcua-modeler==0.4.0
opcua-widgets==0.4.0
openpyxl==2.4.7
Orange3==3.4.1
packaging==16.8
pamela==0.3.0
pandas==0.20.1
pandocfilters==1.4.2
paratext==0.3.1rc1
partd==0.3.8
pathlib2==2.2.1
pathos==0.2.1
patsy==0.4.1
pdfkit==0.6.1
pep8==1.7.0
pexpect==4.2.1
pickleshare==0.7.4
Pillow==4.3.0
pivottablejs==0.8.0
plotly==2.0.15
ply==3.10
pox==0.2.3
ppft==1.6.4.7.1
progressbar2==3.10.1
prompt-toolkit==1.0.14
psutil==5.2.2
psycopg2==2.7.3.1
ptyprocess==0.5.1
py==1.4.33
pycosat==0.6.2
pycparser==2.17
PyCRC==1.21
pycrypto==2.6.1
pycurl==7.43.0
pyelasticsearch==1.4
pyflakes==1.5.0
Pygments==2.2.0
pygpu==0.7.4
pylint==1.6.4
pymc3==3.2
pyodbc==4.0.16
PyOpenGL==3.1.1a1
pyOpenSSL==17.0.0
pyparsing==2.1.4
PyPDF2==1.26.0
pyqtgraph==0.10.0
PySocks==1.6.7
pytest==3.0.7
pytest-runner==2.11.1
python-dateutil==2.6.1
python-editor==1.0.3
python-utils==2.0.0
pytz==2017.2
PyWavelets==0.5.2
PyYAML==3.12
pyzmq==16.0.2
qgrid==0.3.3
QtAwesome==0.4.4
qtconsole==4.3.0
QtPy==1.2.1
requests==2.18.4
rope-py3k==0.9.4.post1
scikit-image==0.13.0
scikit-learn==0.19.0
scipy==0.19.1
seaborn==0.8.1
SecretStorage==2.3.1
simplegeneric==0.8.1
simplejson==3.11.1
singledispatch==3.4.0.3
six==1.11.0
snowballstemmer==1.2.1
sortedcollections==0.5.3
sortedcontainers==1.5.7
Sphinx==1.5.6
spyder==3.1.4
SQLAlchemy==1.1.9
statsmodels==0.8.0
svn==0.3.45
sympy==1.0
tables==3.4.2
tblib==1.3.2
terminado==0.6
testpath==0.3.1
Theano==0.9.0
toolz==0.8.2
tornado==4.5.2
tqdm==4.19.2
traceback2==1.4.0
traitlets==4.3.2
traittypes==0.0.6
txaio==2.8.2
ujson==1.35
unicodecsv==0.14.1
unittest2==1.1.0
urllib3==1.22
vaex==1.0.0b6
vpnotebook==0.1.3
vpython==7.1.3
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.12.2
widgetsnbextension==3.0.3
wikipedia==1.4.0
wrapt==1.10.10
xlrd==1.0.0
XlsxWriter==0.9.6
xlwt==1.2.0
xxhash==1.0.1
zict==0.1.2
